draft desirable methods
There are a reasonable number of different aspects in this commit for
discussion.

Essentially, the regex approach could be simplified to what the
Royal Mail publish as "Valid Postcode Format" [1, p.18] - that
is we would only validate the format and _not_ the data.
+1 from me if we all accept that we are validating the format
alone.

I have also included postcode areas as a resource because frankly
making a regex imo is a needless complexity.

[1]: http://www.royalmail.com/sites/default/files/docs/pdf/programmers_guide_edition_7_v5.pdf